servives & api pagination

+++ QMS_Project/QMS_app/views.py
@@ -1,9 +1,16 @@
 from django.shortcuts import render , redirect
 from django.http import HttpResponse
-from .models import Employee 
-from .forms import EmployeeForm
+from .models import Employee  , Service
+from .forms import EmployeeForm , ServiceForm
 from .filters import EmployeeFilter
 
+
+def dashboard(request):
+    context ={
+
+    }
+
+    return render(request ,"dashboard.html" , context) 
 
 def employees(request):
     emps = Employee.objects.all()
@@ -16,12 +23,11 @@ def employees(request):
     return render(request ,"Employee/employees.html" , context) 
 
 
-def createEmployee(request):
+def addEmployee(request):
 
     form = EmployeeForm()
 
     if request.method == 'POST':
-        print(request.POST)
         form = EmployeeForm(request.POST)
         if form.is_valid():
             form.save()
@@ -43,7 +49,6 @@ def updateEmployee(request, eID):
     form = EmployeeForm(instance = emp)
 
     if request.method == 'POST':
-        print(request.POST)
         form = EmployeeForm(request.POST ,instance = emp)
         if form.is_valid():
             form.save()
@@ -57,9 +62,70 @@ def deleteEmployee(request, eID):
     emp = Employee.objects.get(id = eID)
     context = {"item":emp}
     if request.method == 'POST':
-        emp.delete()
+        if 'Confirm' in request.POST:    
+            emp.delete()
         return redirect('employees')
 
+
+    return render(request ,"delete.html" , context) 
+
+
+
+def services(request):
+    service = Service.objects.all()
+    context = {"serv" : service }
+    return render(request ,"Service/services.html" , context) 
+
+
+
+def addService(request):
+
+    form = ServiceForm()
+
+    if request.method == 'POST':
+        form = ServiceForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('services')
+
+    
+    context ={"form":form}
+    return render(request ,"Service/service_form.html" , context) 
+   
+
+
+def viewService(request, sID):
+
+    context = {}
+    return render(request ,"Service/service.html" , context) 
+
+
+def editService(request, sID):
+
+    
+    ser = Service.objects.get(id = sID)
+    form = ServiceForm(instance = ser)
+
+    if request.method == 'POST':
+        form = ServiceForm(request.POST ,instance = ser)
+        if form.is_valid():
+            form.save()
+            return redirect('services')
+
+    context = {"form":form}
+    return render(request ,"Service/service_form.html" , context)     
+
+
+
+def deleteService(request, sID):
+
+    ser = Service.objects.get(id = sID)
+    context = {"item":ser}
+
+    if request.method == 'POST':
+        if 'Confirm' in request.POST:    
+            ser.delete()
+        return redirect('services')
 
     return render(request ,"delete.html" , context) 
 
@@ -93,9 +159,6 @@ def deleteEmployee(request, eID):
 #     return render(request ,"dashboard.html" , {})    
 
 
-
-def service(request):
-    return render(request ,"services.html" , {})   
 
 
 # def d(request):
